@@ -1,8 +1,8 @@
 apply plugin: 'java-library'
 apply plugin: 'com.novoda.bintray-release'
 
-sourceCompatibility = "1.7"
-targetCompatibility = "1.7"
+sourceCompatibility = JavaVersion.VERSION_1_8
+targetCompatibility = JavaVersion.VERSION_1_8
 
 dependencies {
     implementation fileTree(dir: 'libs', include: ['*.jar'])
@@ -11,6 +11,7 @@ dependencies {
 
     api deps.javapoet
     compileOnly deps.auto.service
+    annotationProcessor deps.auto.service
 }
 
 publish {
