@@ -1,6 +1,6 @@
 package com.youngfeng.snake.demo.annotations;
 
-import android.support.annotation.LayoutRes;
+import androidx.annotation.LayoutRes;
 
 import java.lang.annotation.ElementType;
 import java.lang.annotation.Inherited;
