@@ -2,7 +2,7 @@ package com.youngfeng.snake.app;
 
 import android.animation.Animator;
 import android.os.Bundle;
-import android.support.annotation.Nullable;
+import androidx.annotation.Nullable;
 import android.view.View;
 import android.view.ViewGroup;
 
