@@ -5,8 +5,8 @@ import android.app.Activity;
 import android.app.ActivityOptions;
 import android.content.res.TypedArray;
 import android.os.Build;
-import android.support.annotation.NonNull;
-import android.support.annotation.RequiresApi;
+import androidx.annotation.NonNull;
+import androidx.annotation.RequiresApi;
 
 import java.lang.ref.WeakReference;
 import java.lang.reflect.Method;
