@@ -11,21 +11,21 @@ import android.view.View;
  */
 public class FragmentManagerHelper {
     private FragmentManager mFragmentManager;
-    private android.support.v4.app.FragmentManager mSupportFragmentManager;
+    private androidx.fragment.app.FragmentManager mAndroidXFragmentManager;
 
-    public FragmentManagerHelper(FragmentManager fragmentManager) {
+    private FragmentManagerHelper(FragmentManager fragmentManager) {
         mFragmentManager = fragmentManager;
     }
 
-    public FragmentManagerHelper(android.support.v4.app.FragmentManager fragmentManager) {
-        mSupportFragmentManager = fragmentManager;
+    private FragmentManagerHelper(androidx.fragment.app.FragmentManager fragmentManager) {
+        mAndroidXFragmentManager = fragmentManager;
     }
 
     public static FragmentManagerHelper get(FragmentManager fragmentManager) {
         return new FragmentManagerHelper(fragmentManager);
     }
 
-    public static FragmentManagerHelper get(android.support.v4.app.FragmentManager fragmentManager) {
+    public static FragmentManagerHelper get(androidx.fragment.app.FragmentManager fragmentManager) {
         return new FragmentManagerHelper(fragmentManager);
     }
 
@@ -38,13 +38,13 @@ public class FragmentManagerHelper {
         return mFragmentManager.findFragmentByTag(fragmentTag);
     }
 
-    public android.support.v4.app.Fragment getLastSupportFragment() {
-        int backStackCount = mSupportFragmentManager.getBackStackEntryCount();
+    public androidx.fragment.app.Fragment getLastAndroidXFragment() {
+        int backStackCount = mAndroidXFragmentManager.getBackStackEntryCount();
         if (backStackCount <= 0) return null;
 
-        android.support.v4.app.FragmentManager.BackStackEntry backStackEntry = mSupportFragmentManager.getBackStackEntryAt(backStackCount - 1);
+        androidx.fragment.app.FragmentManager.BackStackEntry backStackEntry = mAndroidXFragmentManager.getBackStackEntryAt(backStackCount - 1);
         String fragmentTag = backStackEntry.getName();
-        return mSupportFragmentManager.findFragmentByTag(fragmentTag);
+        return mAndroidXFragmentManager.findFragmentByTag(fragmentTag);
     }
 
     public View getViewOfLastFragment() {
@@ -52,25 +52,25 @@ public class FragmentManagerHelper {
         return getLastFragment().getView();
     }
 
-    public View getViewOfLastSupportFragment() {
-        if (null == getLastSupportFragment()) return null;
+    public View getViewOfLastAndroidXFragment() {
+        if (null == getLastAndroidXFragment()) return null;
 
-        return getLastSupportFragment().getView();
+        return getLastAndroidXFragment().getView();
     }
 
     public boolean backToLastFragment() {
         return mFragmentManager.popBackStackImmediate();
     }
 
-    public boolean backToSupportFragment() {
-        return mSupportFragmentManager.popBackStackImmediate();
+    public boolean backToAndroidXFragment() {
+        return mAndroidXFragmentManager.popBackStackImmediate();
     }
 
     public boolean backStackEmpty() {
         return mFragmentManager.getBackStackEntryCount() <= 0;
     }
 
-    public boolean supportBackStackEmpty() {
-        return mSupportFragmentManager.getBackStackEntryCount() <= 0;
+    public boolean androidXBackStackEmpty() {
+        return mAndroidXFragmentManager.getBackStackEntryCount() <= 0;
     }
 }
