@@ -1,31 +1,30 @@
 package com.youngfeng.snake.demo.ui;
 
 import android.animation.Animator;
-import android.graphics.Color;
 import android.os.Bundle;
-import android.support.annotation.NonNull;
-import android.support.annotation.Nullable;
-import android.support.v4.app.Fragment;
-import android.support.v7.widget.Toolbar;
 import android.view.LayoutInflater;
 import android.view.View;
 import android.view.ViewGroup;
 import android.widget.LinearLayout;
 import android.widget.TextView;
-
+import androidx.annotation.NonNull;
+import androidx.annotation.Nullable;
+import androidx.appcompat.widget.Toolbar;
+import androidx.fragment.app.Fragment;
+import butterknife.ButterKnife;
 import com.youngfeng.snake.Snake;
 import com.youngfeng.snake.animation.SnakeAnimationController;
+import com.youngfeng.snake.annotations.EnableDragToClose;
 import com.youngfeng.snake.demo.R;
 import com.youngfeng.snake.demo.annotations.BindView;
 import com.youngfeng.snake.demo.ui.widget.TranslateLinearLayout;
-
-import butterknife.ButterKnife;
 
 /**
  * Base class of all support fragments.
  *
  * @author Scott Smith 2017-12-24 10:29
  */
+@EnableDragToClose()
 public class BaseSupportFragment extends Fragment implements SnakeAnimationController {
     private Toolbar mToolbar;
     private TranslateLinearLayout mContentView;
